feat(api): Rename accreditation-serializer to complementary-certification-serializer
